requirements.txt
numpy==1.21.2
pandas==1.3.4
Pillow==9.1.1
scipy==1.7.1
torch==1.9.1
torch_geometric==2.0.1
torchvision==0.2.2
tqdm==4.62.2
